+++ app/src/main/java/com/example/hashcatgui/ApiModels.kt
@@ -3,6 +3,17 @@ package com.example.hashcatgui
 import kotlinx.serialization.Serializable
 
 @Serializable
+data class HashModeInfo(
+    val id: Int,
+    val name: String
+)
+
+@Serializable
+data class HashIdentificationResponse(
+    val hashModes: List<HashModeInfo>
+)
+
+@Serializable
 data class AttackRequest(
     val hash: String,
     val hashType: Int,
