+++ app/src/main/java/com/example/hashcatgui/ui/tabs/AttackTab.kt
@@ -1,0 +1,28 @@
+package com.example.hashcatgui.ui.tabs
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.example.hashcatgui.MainViewModel
+
+@Composable
+fun AttackTab(viewModel: MainViewModel) {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(16.dp),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center
+    ) {
+        Button(onClick = { viewModel.startAttack() }) {
+            Text("Start Remote Attack")
+        }
+    }
+}
